+++ mobile/alix/test/generate_sheet_test.dart
@@ -60,7 +60,7 @@ void main() {
     await tester.pumpWidget(MaterialApp(
       home: PickerScreen(
         root: root,
-        onSetDecksDir: (_) async {},
+        onSetTheme: (_) async {},
         supportDir: support,
         buildClient: buildClient,
         generatePollInterval: const Duration(milliseconds: 10),